Complete getHumanChoice function
Prompt user to type in either 'rock', 'paper', or 'scissors' and added
in a while loop to check their submission. If the user types in anything
other than the three options, the user will be prompted to type in their
choice again until it is one of the three options.